requirements.txt
Flask
pymongo
dnspython
gunicorn
flask-cors
python-dotenv
